Use `NetworkBackup` internally and create a backup manager class
