+++ setup.cfg
@@ -57,3 +57,6 @@ disable_error_code =
     arg-type
 # Only report on selected files
 follow_imports = silent
+
+[tool:pytest]
+asyncio_mode = auto
