+++ zigpy/application.py
@@ -8,11 +8,11 @@ import random
 from typing import Any
 
 import zigpy.appdb
+import zigpy.backup
 import zigpy.config as conf
 import zigpy.device
 import zigpy.exceptions
 import zigpy.group
-import zigpy.models
 import zigpy.ota
 import zigpy.profiles
 import zigpy.quirks
@@ -44,7 +44,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
         self._ota = zigpy.ota.OTA(self)
         self._send_sequence = 0
 
-        self.network_backups: list[zigpy.models.NetworkBackup] = []
+        self.backups: zigpy.backup.BackupManager = zigpy.backup.BackupManager(self)
 
     async def _load_db(self) -> None:
         """Restore save state."""
@@ -55,6 +55,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
         self._dblistener = await zigpy.appdb.PersistingListener.new(database_file, self)
         self.add_listener(self._dblistener)
         self.groups.add_listener(self._dblistener)
+        self.backups.add_listener(self._dblistener)
         await self._dblistener.load()
 
     async def startup(self, *, auto_form: bool = False):
@@ -63,8 +64,6 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
         """
 
         await self.connect()
-
-        formed_new_network = False
 
         try:
             try:
@@ -78,7 +77,6 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
                 LOGGER.info("Forming a new network")
                 await self.form_network()
                 await self.load_network_info(load_devices=False)
-                formed_new_network = True
 
             LOGGER.debug("Network info: %s", self.state.network_info)
             LOGGER.debug("Node info: %s", self.state.node_info)
@@ -92,33 +90,9 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
             await self.shutdown()
             raise
 
-        # Perform a backup in the background
-        asyncio.create_task(
-            self._create_network_backup(formed_new_network=formed_new_network)
+        self.backups.start_periodic_backups(
+            period=self.config[conf.CONF_NWK][zigpy.config.CONF_NWK_BACKUP_PERIOD]
         )
-
-    async def _create_network_backup(self, *, formed_new_network: bool = False) -> None:
-        if formed_new_network:
-            # If we just formed a new network, the backup will contain no devices so we
-            # can just send it to the database listener.
-            self.listener_event(
-                "network_backup_created", self.state.network_info, self.state.node_info
-            )
-            return
-
-        if not self.network_backups or not self.network_backups[-1].compatible_with(
-            network_info=self.state.network_info, node_info=self.state.node_info
-        ):
-            LOGGER.debug(
-                "Loaded state is not compatible with most recent backup: %s",
-                self.network_backups[-1:],
-            )
-            # Do a complete network backup when the superficial network information is
-            # not compatible with the most recent backup present in the database.
-            await self.load_network_info(load_devices=True)
-            self.listener_event(
-                "network_backup_created", self.state.network_info, self.state.node_info
-            )
 
     @classmethod
     async def new(
@@ -208,8 +182,11 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
 
     async def shutdown(self) -> None:
         """Shutdown controller."""
+        self.backups.stop_periodic_backups()
+
         if self._dblistener:
             await self._dblistener.shutdown()
+
         await self.disconnect()
 
     def add_device(self, ieee: t.EUI64, nwk: t.NWK):
