+++ zigpy/config/defaults.py
@@ -1,5 +1,6 @@
 import zigpy.types as t
 
+CONF_NWK_BACKUP_PERIOD_DEFAULT = 24 * 60  # 24 hours
 CONF_NWK_CHANNEL_DEFAULT = 15
 CONF_NWK_CHANNELS_DEFAULT = [15, 20, 25]
 CONF_NWK_EXTENDED_PAN_ID_DEFAULT = None
@@ -14,6 +15,6 @@ CONF_OTA_INOVELLI_DEFAULT = False
 CONF_OTA_LEDVANCE_DEFAULT = False
 CONF_OTA_SALUS_DEFAULT = False
 CONF_OTA_OTAU_DIR_DEFAULT = None
-CONF_TOPO_SCAN_PERIOD_DEFAULT = 240  # 4 hours
+CONF_TOPO_SCAN_PERIOD_DEFAULT = 4 * 60  # 4 hours
 CONF_TOPO_SCAN_ENABLED_DEFAULT = True
 CONF_TOPO_SKIP_COORDINATOR_DEFAULT = False
