+++ zigpy/state.py
@@ -260,204 +260,3 @@ class State:
     @zigpy.util.deprecated("`node_information` has been renamed to `node_info`")
     def node_information(self) -> NodeInfo:
         return self.node_info
-
-
-LOGICAL_TYPE_TO_JSON = {
-    zdo_t.LogicalType.Coordinator: "coordinator",
-    zdo_t.LogicalType.Router: "router",
-    zdo_t.LogicalType.EndDevice: "end_device",
-}
-
-
-JSON_TO_LOGICAL_TYPE = {v: k for k, v in LOGICAL_TYPE_TO_JSON.items()}
-
-
-def network_state_to_json(
-    *, network_info: NetworkInfo, node_info: NodeInfo
-) -> dict[str, Any]:
-    devices = {}
-
-    for ieee, nwk in network_info.nwk_addresses.items():
-        devices[ieee] = {
-            "ieee_address": ieee.serialize()[::-1].hex(),
-            "nwk_address": nwk.serialize()[::-1].hex(),
-            "is_child": False,
-        }
-
-    for ieee in network_info.children:
-        if ieee not in devices:
-            devices[ieee] = {
-                "ieee_address": ieee.serialize()[::-1].hex(),
-                "nwk_address": None,
-                "is_child": True,
-            }
-        else:
-            devices[ieee]["is_child"] = True
-
-    for key in network_info.key_table:
-        if key.partner_ieee not in devices:
-            devices[key.partner_ieee] = {
-                "ieee_address": key.partner_ieee.serialize()[::-1].hex(),
-                "nwk_address": None,
-                "is_child": False,
-            }
-
-        devices[key.partner_ieee]["link_key"] = {
-            "key": key.key.serialize().hex(),
-            "tx_counter": key.tx_counter,
-            "rx_counter": key.rx_counter,
-        }
-
-    return {
-        "metadata": {
-            "version": 1,
-            "format": "zigpy/open-coordinator-backup",
-            "source": network_info.source,
-            "internal": {
-                "node": {
-                    "ieee": node_info.ieee.serialize()[::-1].hex(),
-                    "nwk": node_info.nwk.serialize()[::-1].hex(),
-                    "type": LOGICAL_TYPE_TO_JSON[node_info.logical_type],
-                },
-                "network": {
-                    "tc_link_key": {
-                        "key": network_info.tc_link_key.key.serialize().hex(),
-                        "frame_counter": network_info.tc_link_key.tx_counter,
-                    },
-                    "tc_address": network_info.tc_link_key.partner_ieee.serialize()[
-                        ::-1
-                    ].hex(),
-                    "nwk_manager": network_info.nwk_manager_id.serialize()[::-1].hex(),
-                },
-                "link_key_seqs": {
-                    key.partner_ieee.serialize()[::-1].hex(): key.seq
-                    for key in network_info.key_table
-                },
-                **network_info.metadata,
-            },
-        },
-        "stack_specific": network_info.stack_specific,
-        "coordinator_ieee": node_info.ieee.serialize()[::-1].hex(),
-        "pan_id": network_info.pan_id.serialize()[::-1].hex(),
-        "extended_pan_id": network_info.extended_pan_id.serialize()[::-1].hex(),
-        "nwk_update_id": network_info.nwk_update_id,
-        "security_level": network_info.security_level,
-        "channel": network_info.channel,
-        "channel_mask": list(network_info.channel_mask),
-        "network_key": {
-            "key": network_info.network_key.key.serialize().hex(),
-            "sequence_number": network_info.network_key.seq or 0,
-            "frame_counter": network_info.network_key.tx_counter or 0,
-        },
-        "devices": sorted(devices.values(), key=lambda d: d["ieee_address"]),
-    }
-
-
-def json_to_network_state(obj: dict[str, Any]) -> tuple[NetworkInfo, NodeInfo]:
-    internal = obj["metadata"].get("internal", {})
-
-    node_info = NodeInfo()
-    node_meta = internal.get("node", {})
-
-    if "nwk" in node_meta:
-        node_info.nwk, _ = t.NWK.deserialize(bytes.fromhex(node_meta["nwk"])[::-1])
-    else:
-        node_info.nwk = t.NWK(0x0000)
-
-    node_info.logical_type = JSON_TO_LOGICAL_TYPE[node_meta.get("type", "coordinator")]
-
-    # Should be identical to `metadata.internal.node.ieee`
-    node_info.ieee, _ = t.EUI64.deserialize(
-        bytes.fromhex(obj["coordinator_ieee"])[::-1]
-    )
-
-    network_info = NetworkInfo()
-    network_info.source = obj["metadata"]["source"]
-    network_info.metadata = {
-        k: v
-        for k, v in internal.items()
-        if k not in ("node", "network", "link_key_seqs")
-    }
-    network_info.pan_id, _ = t.NWK.deserialize(bytes.fromhex(obj["pan_id"])[::-1])
-    network_info.extended_pan_id, _ = t.EUI64.deserialize(
-        bytes.fromhex(obj["extended_pan_id"])[::-1]
-    )
-    network_info.nwk_update_id = obj["nwk_update_id"]
-
-    network_meta = internal.get("network", {})
-
-    if "nwk_manager" in network_meta:
-        network_info.nwk_manager_id, _ = t.NWK.deserialize(
-            bytes.fromhex(network_meta["nwk_manager"])
-        )
-    else:
-        network_info.nwk_manager_id = t.NWK(0x0000)
-
-    network_info.channel = obj["channel"]
-    network_info.channel_mask = t.Channels.from_channel_list(obj["channel_mask"])
-    network_info.security_level = obj["security_level"]
-
-    if obj.get("stack_specific"):
-        network_info.stack_specific = obj.get("stack_specific")
-
-    network_info.tc_link_key = Key()
-
-    if "tc_link_key" in network_meta:
-        network_info.tc_link_key.key, _ = t.KeyData.deserialize(
-            bytes.fromhex(network_meta["tc_link_key"]["key"])
-        )
-        network_info.tc_link_key.tx_counter = network_meta["tc_link_key"].get(
-            "frame_counter", 0
-        )
-        network_info.tc_link_key.partner_ieee, _ = t.EUI64.deserialize(
-            bytes.fromhex(network_meta["tc_address"])[::-1]
-        )
-    else:
-        network_info.tc_link_key.key = conf.CONF_NWK_TC_LINK_KEY_DEFAULT
-        network_info.tc_link_key.partner_ieee = node_info.ieee
-
-    network_info.network_key = Key()
-    network_info.network_key.key, _ = t.KeyData.deserialize(
-        bytes.fromhex(obj["network_key"]["key"])
-    )
-    network_info.network_key.tx_counter = obj["network_key"]["frame_counter"]
-    network_info.network_key.seq = obj["network_key"]["sequence_number"]
-
-    network_info.children = []
-    network_info.nwk_addresses = {}
-
-    for device in obj["devices"]:
-        if device["nwk_address"] is not None:
-            nwk, _ = t.NWK.deserialize(bytes.fromhex(device["nwk_address"])[::-1])
-        else:
-            nwk = None
-
-        ieee, _ = t.EUI64.deserialize(bytes.fromhex(device["ieee_address"])[::-1])
-
-        # The `is_child` key is currently optional
-        if device.get("is_child", True):
-            network_info.children.append(ieee)
-
-        if nwk is not None:
-            network_info.nwk_addresses[ieee] = nwk
-
-        if "link_key" in device:
-            key = Key()
-            key.key, _ = t.KeyData.deserialize(bytes.fromhex(device["link_key"]["key"]))
-            key.tx_counter = device["link_key"]["tx_counter"]
-            key.rx_counter = device["link_key"]["rx_counter"]
-            key.partner_ieee = ieee
-
-            try:
-                key.seq = obj["metadata"]["internal"]["link_key_seqs"][
-                    device["ieee_address"]
-                ]
-            except KeyError:
-                key.seq = 0
-
-            network_info.key_table.append(key)
-
-        # XXX: Devices that are not children, have no NWK address, and have no link key
-        #      are effectively ignored, since there is no place to write them
-
-    return network_info, node_info
